[add] made getWeekData more modular
