+++ db/config.js
@@ -4,7 +4,6 @@
 const pg = require('pg');
 const Sequelize = require('sequelize');
 
-//'postgres://localhost/testyourstory'
 const db = new Sequelize(process.env.DATABASE_URL || 'postgres://localhost/testyourstory', {
   dialect: 'postgres',
   define: {
