+++ package.json
@@ -7,7 +7,7 @@
     "start": "nodemon ./server/server.js",
     "build": "webpack -p --config webpack.config.production.js",
     "build:dev": "webpack --watch --config webpack.config.development.js",
-    "test": "jasmine"
+    "test": "NODE_ENV='postgres://localhost/testyourstory' jasmine"
   },
   "repository": {
     "type": "git",
